app configuration support for more properties

+++ internal/services/appconfiguration/app_configuration_resource.go
@@ -3,6 +3,7 @@ package appconfiguration
 import (
 	"fmt"
 	"log"
+	"net/http"
 	"strings"
 	"time"
 
@@ -12,6 +13,7 @@ import (
 	"github.com/hashicorp/go-azure-helpers/resourcemanager/location"
 	"github.com/hashicorp/go-azure-helpers/resourcemanager/tags"
 	"github.com/hashicorp/go-azure-sdk/resource-manager/appconfiguration/2022-05-01/configurationstores"
+	"github.com/hashicorp/go-azure-sdk/resource-manager/appconfiguration/2022-05-01/deletedconfigurationstores"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/azure"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/tf"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
@@ -19,6 +21,7 @@ import (
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/validation"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/timeouts"
+	"github.com/hashicorp/terraform-provider-azurerm/utils"
 )
 
 func resourceAppConfiguration() *pluginsdk.Resource {
@@ -52,7 +55,44 @@ func resourceAppConfiguration() *pluginsdk.Resource {
 
 			"resource_group_name": azure.SchemaResourceGroupName(),
 
+			"encryption": {
+				Type:     pluginsdk.TypeList,
+				Optional: true,
+				MaxItems: 1,
+				Elem: &pluginsdk.Resource{
+					Schema: map[string]*pluginsdk.Schema{
+						"key_vault_key_identifier": {
+							Type:         pluginsdk.TypeString,
+							Optional:     true,
+							ValidateFunc: validation.IsURLWithHTTPorHTTPS,
+						},
+						"identity_client_id": {
+							Type:         pluginsdk.TypeString,
+							Optional:     true,
+							ValidateFunc: validation.IsUUID,
+						},
+					},
+				},
+			},
+
 			"identity": commonschema.SystemAssignedUserAssignedIdentityOptional(),
+
+			"local_auth_enabled": {
+				Type:     pluginsdk.TypeBool,
+				Optional: true,
+				Default:  true,
+			},
+
+			"public_network_access_enabled": {
+				Type:     pluginsdk.TypeBool,
+				Optional: true,
+			},
+
+			"purge_protection_enabled": {
+				Type:     pluginsdk.TypeBool,
+				Optional: true,
+				Default:  false,
+			},
 
 			"sku": {
 				Type:     pluginsdk.TypeString,
@@ -62,6 +102,14 @@ func resourceAppConfiguration() *pluginsdk.Resource {
 					"free",
 					"standard",
 				}, false),
+			},
+
+			"soft_delete_retention_days": {
+				Type:         pluginsdk.TypeInt,
+				Optional:     true,
+				Default:      7,
+				ForceNew:     true,
+				ValidateFunc: validation.IntBetween(1, 7),
 			},
 
 			"endpoint": {
@@ -172,6 +220,7 @@ func resourceAppConfiguration() *pluginsdk.Resource {
 
 func resourceAppConfigurationCreate(d *pluginsdk.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).AppConfiguration.ConfigurationStoresClient
+	deletedConfigurationStoresClient := meta.(*clients.Client).AppConfiguration.DeletedConfigurationStoresClient
 	subscriptionId := meta.(*clients.Client).Account.SubscriptionId
 	ctx, cancel := timeouts.ForCreate(meta.(*clients.Client).StopContext, d)
 	defer cancel()
@@ -191,12 +240,51 @@ func resourceAppConfigurationCreate(d *pluginsdk.ResourceData, meta interface{})
 		return tf.ImportAsExistsError("azurerm_app_configuration", resourceId.ID())
 	}
 
+	location := azure.NormalizeLocation(d.Get("location").(string))
+	deletedConfigurationStoresId := deletedconfigurationstores.NewDeletedConfigurationStoreID(subscriptionId, location, name)
+	deleted, err := deletedConfigurationStoresClient.ConfigurationStoresGetDeleted(ctx, deletedConfigurationStoresId)
+	if err != nil {
+		if !response.WasNotFound(deleted.HttpResponse) {
+			return fmt.Errorf("checking for presence of deleted %s: %+v", deletedConfigurationStoresId, name)
+		}
+	}
+
+	recoverSoftDeleted := false
+	if !response.WasNotFound(deleted.HttpResponse) && !response.WasStatusCode(deleted.HttpResponse, http.StatusForbidden) {
+		if !meta.(*clients.Client).Features.AppConfiguration.RecoverSoftDeleted {
+			return fmt.Errorf(optedOutOfRecoveringSoftDeletedAppConfigurationErrorFmt(name, location))
+		}
+		recoverSoftDeleted = true
+	}
+
 	parameters := configurationstores.ConfigurationStore{
-		Location: azure.NormalizeLocation(d.Get("location").(string)),
+		Location: location,
 		Sku: configurationstores.Sku{
 			Name: d.Get("sku").(string),
 		},
+		Properties: &configurationstores.ConfigurationStoreProperties{
+			EnablePurgeProtection: utils.Bool(d.Get("purge_protection_enabled").(bool)),
+			DisableLocalAuth:      utils.Bool(!d.Get("local_auth_enabled").(bool)),
+			Encryption:            expandAppConfigurationEncryption(d.Get("encryption").([]interface{})),
+		},
 		Tags: tags.Expand(d.Get("tags").(map[string]interface{})),
+	}
+
+	if v, ok := d.Get("soft_delete_retention_days").(int); ok && v != 7 {
+		parameters.Properties.SoftDeleteRetentionInDays = utils.Int64(int64(v))
+	}
+
+	if v := d.GetRawConfig().AsValueMap()["public_network_access_enabled"]; !v.IsNull() {
+		publicNetworkAccess := configurationstores.PublicNetworkAccessEnabled
+		if v.False() {
+			publicNetworkAccess = configurationstores.PublicNetworkAccessDisabled
+		}
+		parameters.Properties.PublicNetworkAccess = &publicNetworkAccess
+	}
+
+	if recoverSoftDeleted {
+		t := configurationstores.CreateModeRecover
+		parameters.Properties.CreateMode = &t
 	}
 
 	identity, err := identity.ExpandSystemAndUserAssignedMap(d.Get("identity").([]interface{}))
@@ -224,11 +312,28 @@ func resourceAppConfigurationUpdate(d *pluginsdk.ResourceData, meta interface{})
 		return err
 	}
 
-	parameters := configurationstores.ConfigurationStoreUpdateParameters{
-		Sku: &configurationstores.Sku{
+	existing, err := client.Get(ctx, *id)
+	if err != nil {
+		return fmt.Errorf("retrieving %s: %+v", *id, err)
+	}
+	if existing.Model == nil {
+		return fmt.Errorf("retrieving %s: `model` was nil", *id)
+	}
+	if existing.Model.Properties == nil {
+		return fmt.Errorf("retrieving %s: `properties` was nil", *id)
+	}
+
+	update := configurationstores.ConfigurationStoreUpdateParameters{}
+
+	if d.HasChange("sku") {
+		update.Sku = &configurationstores.Sku{
 			Name: d.Get("sku").(string),
-		},
-		Tags: tags.Expand(d.Get("tags").(map[string]interface{})),
+		}
+	}
+
+	if d.HasChange("tags") {
+		t := d.Get("tags").(map[string]interface{})
+		update.Tags = tags.Expand(t)
 	}
 
 	if d.HasChange("identity") {
@@ -236,10 +341,58 @@ func resourceAppConfigurationUpdate(d *pluginsdk.ResourceData, meta interface{})
 		if err != nil {
 			return fmt.Errorf("expanding `identity`: %+v", err)
 		}
-		parameters.Identity = identity
+		update.Identity = identity
 	}
 
-	if err := client.UpdateThenPoll(ctx, *id, parameters); err != nil {
+	if d.HasChange("encryption") {
+		if update.Properties == nil {
+			update.Properties = &configurationstores.ConfigurationStorePropertiesUpdateParameters{}
+		}
+		update.Properties.Encryption = expandAppConfigurationEncryption(d.Get("encryption").([]interface{}))
+	}
+
+	if d.HasChange("local_auth_enabled") {
+		if update.Properties == nil {
+			update.Properties = &configurationstores.ConfigurationStorePropertiesUpdateParameters{}
+		}
+		update.Properties.DisableLocalAuth = utils.Bool(!d.Get("local_auth_enabled").(bool))
+	}
+
+	if d.HasChange("purge_protection_enabled") {
+		if update.Properties == nil {
+			update.Properties = &configurationstores.ConfigurationStorePropertiesUpdateParameters{}
+		}
+		newValue := d.Get("purge_protection_enabled").(bool)
+
+		oldValue := false
+		if existing.Model.Properties.EnablePurgeProtection != nil {
+			oldValue = *existing.Model.Properties.EnablePurgeProtection
+		}
+
+		if oldValue && !newValue {
+			return fmt.Errorf("updating %s: once Purge Protection has been Enabled it's not possible to disable it", *id)
+		}
+		update.Properties.EnablePurgeProtection = utils.Bool(d.Get("purge_protection_enabled").(bool))
+	}
+
+	if d.HasChange("public_network_enabled") {
+		v := d.GetRawConfig().AsValueMap()["public_network_access_enabled"]
+		if v.IsNull() && existing.Model.Properties.SoftDeleteRetentionInDays != nil {
+			return fmt.Errorf("updating %s: once Public Network Access has been explicitly Enabled or Disabled it's not possible to unset it to which means Automatic", *id)
+		}
+
+		if update.Properties == nil {
+			update.Properties = &configurationstores.ConfigurationStorePropertiesUpdateParameters{}
+		}
+
+		publicNetworkAccess := configurationstores.PublicNetworkAccessEnabled
+		if v.False() {
+			publicNetworkAccess = configurationstores.PublicNetworkAccessDisabled
+		}
+		update.Properties.PublicNetworkAccess = &publicNetworkAccess
+	}
+
+	if err := client.UpdateThenPoll(ctx, *id, update); err != nil {
 		return fmt.Errorf("updating %s: %+v", *id, err)
 	}
 
@@ -280,6 +433,30 @@ func resourceAppConfigurationRead(d *pluginsdk.ResourceData, meta interface{}) e
 
 		if props := model.Properties; props != nil {
 			d.Set("endpoint", props.Endpoint)
+			d.Set("encryption", flattenAppConfigurationEncryption(props.Encryption))
+			d.Set("public_network_access_enabled", flattenAppConfigurationPublicNetworkAccess(props.PublicNetworkAccess))
+
+			localAuthEnabled := true
+			if props.DisableLocalAuth != nil {
+				localAuthEnabled = !(*props.DisableLocalAuth)
+			}
+
+			d.Set("local_auth_enabled", localAuthEnabled)
+
+			purgeProtectionEnabled := false
+			if props.EnablePurgeProtection != nil {
+				purgeProtectionEnabled = *props.EnablePurgeProtection
+			}
+			d.Set("purge_protection_enabled", purgeProtectionEnabled)
+
+			// In 2022-05-01 version API, for standard sku, service will return soft_delete_retention_days=7,
+			// even if it is not set in PUT method. But for free sku, service will return
+			// soft_delete_retention_days=0. To be consistent, here we set 0 values to 7.
+			softDeleteRetentionDays := 7
+			if props.SoftDeleteRetentionInDays != nil && *props.SoftDeleteRetentionInDays != 0 {
+				softDeleteRetentionDays = int(*props.SoftDeleteRetentionInDays)
+			}
+			d.Set("soft_delete_retention_days", softDeleteRetentionDays)
 		}
 
 		accessKeys := flattenAppConfigurationAccessKeys(resultPage.Items)
@@ -304,7 +481,9 @@ func resourceAppConfigurationRead(d *pluginsdk.ResourceData, meta interface{}) e
 
 func resourceAppConfigurationDelete(d *pluginsdk.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).AppConfiguration.ConfigurationStoresClient
+	deletedConfigurationStoresClient := meta.(*clients.Client).AppConfiguration.DeletedConfigurationStoresClient
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
+	subscriptionId := meta.(*clients.Client).Account.SubscriptionId
 	defer cancel()
 
 	id, err := configurationstores.ParseConfigurationStoreID(d.Id())
@@ -312,8 +491,65 @@ func resourceAppConfigurationDelete(d *pluginsdk.ResourceData, meta interface{})
 		return err
 	}
 
+	existing, err := client.Get(ctx, *id)
+	if err != nil {
+		if !response.WasNotFound(existing.HttpResponse) {
+			return nil
+		}
+		return fmt.Errorf("retrieving %s: %+v", *id, err)
+	}
+
+	if existing.Model == nil {
+		return fmt.Errorf("retrieving %q: `model` was nil", *id)
+	}
+	if existing.Model.Properties == nil {
+		return fmt.Errorf("retrieving %q: `properties` was nil", *id)
+	}
+
+	purgeProtectionEnabled := false
+	if ppe := existing.Model.Properties.EnablePurgeProtection; ppe != nil {
+		purgeProtectionEnabled = *ppe
+	}
+	softDeleteEnabled := false
+	if sde := existing.Model.Properties.SoftDeleteRetentionInDays; sde != nil && *sde > 0 {
+		softDeleteEnabled = true
+	}
+
 	if err := client.DeleteThenPoll(ctx, *id); err != nil {
 		return fmt.Errorf("deleting %s: %+v", *id, err)
+	}
+
+	if meta.(*clients.Client).Features.AppConfiguration.PurgeSoftDeleteOnDestroy && softDeleteEnabled {
+		deletedId := deletedconfigurationstores.NewDeletedConfigurationStoreID(subscriptionId, existing.Model.Location, id.ConfigStoreName)
+
+		// AppConfiguration with Purge Protection Enabled cannot be deleted unless done by Azure
+		if purgeProtectionEnabled {
+			deletedInfo, err := deletedConfigurationStoresClient.ConfigurationStoresGetDeleted(ctx, deletedId)
+			if err != nil {
+				return fmt.Errorf("retrieving the Deletion Details for %s: %+v", *id, err)
+			}
+
+			if deletedInfo.Model != nil && deletedInfo.Model.Properties != nil && deletedInfo.Model.Properties.DeletionDate != nil && deletedInfo.Model.Properties.ScheduledPurgeDate != nil {
+				log.Printf("[DEBUG] The App Configuration %q has Purge Protection Enabled and was deleted on %q. Azure will purge this on %q",
+					id.ConfigStoreName, *deletedInfo.Model.Properties.DeletionDate, *deletedInfo.Model.Properties.ScheduledPurgeDate)
+			} else {
+				log.Printf("[DEBUG] The App Configuration %q has Purge Protection Enabled and will be purged automatically by Azure", id.ConfigStoreName)
+			}
+			return nil
+		}
+
+		log.Printf("[DEBUG]  %q marked for purge - executing purge", id.ConfigStoreName)
+		future, err := deletedConfigurationStoresClient.ConfigurationStoresPurgeDeleted(ctx, deletedId)
+		if err != nil {
+			return err
+		}
+
+		log.Printf("[DEBUG] Waiting for purge of App Configuration %q..", id.ConfigStoreName)
+
+		if err := future.Poller.PollUntilDone(); err != nil {
+			return fmt.Errorf("waiting for purging of %s: %+v", *id, err)
+		}
+		log.Printf("[DEBUG] Purged AppConfiguration %q.", id.ConfigStoreName)
 	}
 
 	return nil
@@ -324,6 +560,25 @@ type flattenedAccessKeys struct {
 	primaryWriteKey   []interface{}
 	secondaryReadKey  []interface{}
 	secondaryWriteKey []interface{}
+}
+
+func expandAppConfigurationEncryption(input []interface{}) *configurationstores.EncryptionProperties {
+	if len(input) == 0 {
+		return nil
+	}
+
+	encryptionParam := input[0].(map[string]interface{})
+	result := &configurationstores.EncryptionProperties{
+		KeyVaultProperties: &configurationstores.KeyVaultProperties{},
+	}
+
+	if v, ok := encryptionParam["identity_client_id"].(string); ok && v != "" {
+		result.KeyVaultProperties.IdentityClientId = &v
+	}
+	if v, ok := encryptionParam["key_vault_key_identifier"].(string); ok && v != "" {
+		result.KeyVaultProperties.KeyIdentifier = &v
+	}
+	return result
 }
 
 func flattenAppConfigurationAccessKeys(values []configurationstores.ApiKey) flattenedAccessKeys {
@@ -387,4 +642,20 @@ func flattenAppConfigurationAccessKey(input configurationstores.ApiKey) []interf
 			"secret":            secret,
 		},
 	}
+}
+
+func optedOutOfRecoveringSoftDeletedAppConfigurationErrorFmt(name, location string) string {
+	return fmt.Sprintf(`
+An existing soft-deleted App Configuration exists with the Name %q in the location %q, however
+automatically recovering this App Configuration has been disabled via the "features" block.
+
+Terraform can automatically recover the soft-deleted App Configuration when this behaviour is
+enabled within the "features" block (located within the "provider" block) - more information
+can be found here:
+
+https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs#features
+
+Alternatively you can manually recover this (e.g. using the Azure CLI) and then import
+this into Terraform via "terraform import", or pick a different name/location.
+`, name, location)
 }
